+++ lcviz/plugins/flux_column/__init__.py
@@ -1,0 +1,1 @@
+from .flux_column import *  # noqa